requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
jwt==1.3.1
polygon-api-client==1.13.7
pycparser==2.22
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.18
websockets==12.0